Added figure rendering

+++ include/renderer.hpp
@@ -8,6 +8,6 @@ namespace Renderer
 
     void line_width(float width);
     void render_background();
-    void draw_line(glm::ivec2 point1, glm::ivec2 point2, const glm::vec3& color1 = glm::vec3(1.0),
+    void draw_line(glm::vec2 point1, glm::vec2 point2, const glm::vec3& color1 = glm::vec3(1.0),
                    const glm::vec3& color2 = glm::vec3(1.0), bool can_use_props_colors = true);
 }// namespace Renderer
